Fix: vm.h에서 enum vm_type 관련 코드 제거 및 uninit.h로 이동, uninit.h에서 vm_type.h 포함. vm.c에서 frame_table 전역 변수 추가 및 vm_dealloc_frame 함수 구현. anon.c에서 swap_disk 변수 제거.
